+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add ingest and delete datapoints capability
+
 ## [0.25.1] - 2024-09-12
 
 ### Added